feat: add a map legend
