+++ covid-dashboard/src/js/html/appHtml.html
@@ -6,11 +6,52 @@
     <div class="app__col1">
       <div class="table2">
         Table 2
+        <button id="testTooltip">Tooltip </button>
       </div>
     </div>
     <div class="app__col2">
       <div class="map" id="leafletMap">
-        Map
+        <div class="legend">
+          <p class="legend__title">Map legend <span class="legend__arrow">›</span></p>
+          <div class="legend__inner">
+            <div class="legend__menu">
+              <div class="legend__period">
+                <input type="radio" checked name="mapPeriod" id="mapLastDay">
+                <label for="mapLastDay">
+                  Last day
+                </label>
+                <input type="radio" name="mapPeriod" id="mapAllTime">
+                <label for="mapAllTime">
+                  All time
+                </label>
+              </div>
+              <div class="legend__quantity">
+                <input type="radio" checked name="mapQuantity" id="mapAllPopulation">
+                <label for="mapAllPopulation">
+                  All population
+                </label>
+                <input type="radio" name="mapQuantity" id="mapPer100k">
+                <label for="mapPer100k">
+                  Per 100K
+                </label>
+              </div>
+            </div>
+            <div class="infection">
+              <p class="infection__title">Infection levels</p>
+              <div class="infection__box">
+                <div class="infection__circle">
+                  high
+                </div>
+                <div class="infection__arrow">
+                  -› -›
+                </div>
+                <div class="infection__circle infection__circle_small">
+                  low
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
       </div>
     </div>
     <div class="app__col3">
