+++ covid-dashboard/src/js/modules/mapLegendInit.js
@@ -1,0 +1,8 @@
+import dom from './dom.js';
+
+export default () => {
+  dom.legendTitle.addEventListener('click', () => {
+    dom.legendInner.classList.toggle('legend__inner_active');
+    dom.legendArrow.classList.toggle('legend__arrow_active');
+  });
+};
