+++ covid-dashboard/src/styles/_app.scss
@@ -9,6 +9,8 @@
   }
 
   &__header {
+    position: relative;
+    z-index: 999;
     font-weight: 900;
     font-size: 1.5rem;
     padding-bottom: 0.5rem;
@@ -30,11 +32,14 @@
   }
 
   &__col1 {
-    flex-basis: 20%;
+    position: relative;
+    z-index: 999;
+    width: 270px;
     margin-right: 1rem;
 
     @media (max-width: $screen-laptop) {
       height: 30vh;
+      width: auto;
       margin-bottom: 1rem;
       margin-right: 0;
     }
@@ -45,6 +50,7 @@
     margin-right: 1rem;
 
     @media (max-width: $screen-laptop) {
+      position: relative;
       height: 30vh;
       margin-bottom: 1rem;
       margin-right: 0;
@@ -52,16 +58,22 @@
   }
 
   &__col3 {
+    position: relative;
+    z-index: 999;
     display: flex;
     flex-direction: column;
-    flex-basis: 30%;
+    justify-content: space-between;
+    width: 400px;
 
     @media (max-width: $screen-laptop) {
       height: auto;
+      width: auto;
     }
   }
 
   &__footer {
+    position: relative;
+    z-index: 999;
     padding-top: 1rem;
   }
