+++ covid-dashboard/src/styles/_legend.scss
@@ -1,0 +1,68 @@
+.legend {
+  position: absolute;
+  right: 430px;
+  bottom: 52px;
+  z-index: 999;
+  font-family: 'Fira Sans', sans-serif;
+  background-color: rgba(#fff, 0.9);
+  border-radius: 5px;
+  padding: 1rem;
+
+  &__inner {
+    height: 0px;
+    overflow: hidden;
+    transition: 0.3s;
+
+    &_active {
+      height: auto;
+    }
+  }
+
+  &__title {
+    font-weight: 600;
+    font-size: 1.2rem;
+    cursor: pointer;
+    transition: 0.3s;
+
+    &:hover {
+      opacity: 0.7;
+    }
+
+    &:hover .legend__arrow {
+      transform: rotate(270deg);
+    }
+
+    &:hover .legend__arrow_active {
+      transform: rotate(90deg);
+    }
+  }
+
+  &__arrow {
+    display: inline-block;
+    font-weight: 600;
+    transform-origin: center;
+    transform: rotate(90deg);
+    margin-left: 0.5rem;
+    transition: 0.3s;
+
+    &_active {
+      transform: rotate(270deg);
+    }
+  }
+
+  &__menu {
+    display: flex;
+    gap: 1rem;
+    justify-content: space-between;
+    font-size: 0.8rem;
+    margin-top: 0.5rem;
+  }
+
+  &__period, &__quantity {
+    border: 1px solid #ccc;
+    border-radius: 0.3rem;
+    padding: 0.3rem;
+  }
+
+
+}
