+++ covid-dashboard/src/styles/_map.scss
@@ -1,5 +1,13 @@
 .map {
-  border: 1px solid #ccc;
-  height: 100%;
-  padding: 1rem;
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: 0;
+  bottom: 0;
+
+  @media (max-width: $screen-laptop) {
+    position: static;
+    height: 100%;
+  }
+
 }
