+++ covid-dashboard/src/styles/_table1.scss
@@ -1,14 +1,8 @@
 .table1 {
-  display: flex;
-  flex-direction: column;
-  height: calc(50vh - 1rem);
-  border: 1px solid #ccc;
+  background-color: rgba(#fff, 0.9);
+  border-radius: 5px;
   margin-bottom: 1rem;
   padding: 1rem;
-
-  @media (max-width: $screen-laptop) {
-    height: 50vh;
-  }
 
   &__menu {
     display: flex;
