+++ covid-dashboard/src/styles/_table2.scss
@@ -1,5 +1,6 @@
 .table2 {
   height: 100%;
-  border: 1px solid #ccc;
   padding: 1rem;
+  background-color: rgba(#fff, 0.9);
+  border-radius: 5px;;
 }
