+++ covid-dashboard/src/styles/index.scss
@@ -8,3 +8,5 @@
 @import "_table2.scss";
 @import "_switch-block.scss";
 @import "_search.scss";
+@import "_legend.scss";
+@import "_infection.scss";
