+++ covid-dashboard/src/styles/leaflet_add.scss
@@ -1,0 +1,4 @@
+.leaflet-control-zoom.leaflet-bar.leaflet-control {
+  display: flex;
+  gap: 2px;
+}
